+++ RoSA_Monitor/launchermanager.cpp
@@ -94,7 +94,7 @@ void LauncherManager::StopLauncher(LauncherType type){
              auxCommand = "cd && cd " + workspace_dir->path() + " && source install/setup.bash && ros2 run teleop_twist_qt teleop_twist_qt";
              break;
          case FIRMWARE:
-             auxCommand = "cd && cd " + workspace_dir->path() + " && source install/setup.bash && rqt";
+             auxCommand = "cd && cd " + workspace_dir->path() + " && source install/setup.bash && ros2 run rosa_firmware rosa_driver --ros-args -p device_name:=/dev/ttyUSB0";
              break;
          case GAZEBO_SIM:
              auxCommand = "cd && cd " + workspace_dir->path() + " && source install/setup.bash && ros2 launch rosa_description rosa_gazebo_launch.py use_sim_time:=true";