+++ RoSA_Monitor/simulationwindow.h
@@ -8,6 +8,7 @@
 #include <QFileDialog>
 
 #include "launchermanager.h"
+#include "shelloutputwindow.h"
 
 namespace Ui {
 class SimulationWindow;
@@ -78,10 +79,21 @@ protected:
     //Set buttons state to off when window is closed
     void shutDownWindow();
 
+    //Delete a shell tab and reorganize index
+    void RemoveTab(LauncherManager::LauncherType type);
+
 private:
     Ui::SimulationWindow *ui;
 
     LauncherManager manager;
+
+    int numOfTabs;
+
+    //Index assigned for each active tab; 0 if not active
+    int tab_Index[LauncherManager::NUMBER_OF_LAUNCHERS];
+
+    //Array of shellWindows objects for independent tabs
+    ShellOutputWindow* shellWindows[LauncherManager::NUMBER_OF_LAUNCHERS];
 
 
 };